Update to latest OpenAPI spec

+++ typescript/datakatalog-kubb/schemas/assosiasjonstypeSchema.ts
@@ -3,11 +3,11 @@ import { assosiasjonstypeVegobjekttypeSchema } from './assosiasjonstypeVegobjekt
 import { z } from 'zod'
 
 export const assosiasjonstypeSchema = z.object({
-  id: z.number().int(),
+  id: z.number(),
   navn: z.string().optional(),
   egenskapstype: z.string(),
-  maksimalt_antall_verdier: z.number().int().optional(),
-  minimalt_antall_verdier: z.number().int().optional(),
+  maksimalt_antall_verdier: z.number().optional(),
+  minimalt_antall_verdier: z.number().optional(),
   innhold: z.lazy(() => assosiasjonstypeListeInnholdSchema).optional(),
   type: z.lazy(() => assosiasjonstypeVegobjekttypeSchema).optional(),
 })
