+++ typescript/datakatalog-kubb/schemas/egenskapstypeKategoriSchema.ts
@@ -1,9 +1,9 @@
 import { z } from 'zod'
 
 export const egenskapstypeKategoriSchema = z.object({
-  id: z.number().int(),
+  id: z.number(),
   navn: z.string().optional(),
   kortnavn: z.string().optional(),
-  sorteringsnummer: z.number().int(),
+  sorteringsnummer: z.number(),
   beskrivelse: z.string().optional(),
 })
