+++ typescript/datakatalog-kubb/schemas/enhetSchema.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod'
 
 export const enhetSchema = z.object({
-  id: z.number().int(),
+  id: z.number(),
   navn: z.string().optional(),
   kortnavn: z.string().optional(),
 })
