+++ typescript/datakatalog-kubb/schemas/stedfestingListeSchema.ts
@@ -6,8 +6,8 @@ export const stedfestingListeSchema = z
   .lazy(() => stedfestingSchema)
   .and(
     z.object({
-      maksimalt_antall_verdier: z.number().int().optional(),
-      minimalt_antall_verdier: z.number().int().optional(),
+      maksimalt_antall_verdier: z.number().optional(),
+      minimalt_antall_verdier: z.number().optional(),
       innhold: z.lazy(() => egenskapstypeStedfestingSchema).optional(),
     }),
   )
