+++ typescript/vegobjekter/schemas/riksvegruteSchema.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
 
 export const riksvegruteSchema = z.object({
-  enumid: z.number().int(),
+  enumid: z.number(),
   riksvegrute: z.string(),
 })
