+++ typescript/vegobjekter/schemas/tekstEnumEgenskapSchema.ts
@@ -4,8 +4,5 @@ import { z } from 'zod'
 export const tekstEnumEgenskapSchema = z
   .lazy(() => egenskapSchema)
   .and(
-    z.object({
-      verdi: z.string().optional(),
-      enum_id: z.number().int().optional(),
-    }),
+    z.object({ verdi: z.string().optional(), enum_id: z.number().optional() }),
   )
