+++ typescript/vegobjekter/schemas/vegforvalterSchema.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
 
 export const vegforvalterSchema = z.object({
-  enumid: z.number().int(),
+  enumid: z.number(),
   vegforvalter: z.string(),
 })
