+++ typescript/vegobjekter/schemas/vegobjektTypeSchema.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
 
 export const vegobjektTypeSchema = z.object({
-  id: z.number().int(),
+  id: z.number(),
   navn: z.string(),
 })
