+++ typescript/vegobjekter/schemas/vegobjekterStatistikkMedTypeSchema.ts
@@ -2,7 +2,7 @@ import { vegobjekterStatistikkSchema } from './vegobjekterStatistikkSchema'
 import { z } from 'zod'
 
 export const vegobjekterStatistikkMedTypeSchema = z.object({
-  id: z.number().int(),
+  id: z.number(),
   navn: z.string(),
   href: z.string(),
   statistikk: z.lazy(() => vegobjekterStatistikkSchema),
